Add split of /proc/interrupts text data
The interrupt data collection does a 'cat /proc/interrupts' every
interval until n seconds have elapsed.
This data has each cpu as a column and each interrupt as a row.
Currently I sum each interrupt across the cpus and get a total
for each interrupt per seecond.
I display the change in the number of interrupts.
If an interrupt is all zero over the data collection then I skip
printing it.
I also sum all the interrupts for an interval and add a 'total'
column.
This makes a more manageable chart.
But it skips showing the per cpu nature of the interrupts.
This can hide an imbalance in the interrupts (say all the
interrupts occuring on 1 cpu versus being spread out.