Add cmdline getopts, output filename and images
Currently args to tsv_2_xlsx.py are all positional and only the
tsv filenames are allowed.
Add getopts cmdline processing to allow (optionally) setting the
output file name and (optionally) specifying images to be inserted
into a 'images' sheet.
The images can be specified like '-i "*.png". The glob will be
expanded by python. If you don't use the dbl quote of the glob
then the OS will expand the glob and (if there is more than 1
image file) only the 1st image filename will be passed to the -i
option. The subsequent image filenames will be passed as to the
tsv file processing logic... not good. So be sure to enclose your
image glob in "".