Test if values are numbers and convert to float if so
Currently when I write a row of data, xlsxwrite looks at the type
of each data[] element and writes the value according to the type.
If I don't chart a column of the data (say the 1st col in the
chart is not col 0) then xlsxwriter will write it as a string...
which is annoying at excel puts all those stupid corner notes on
each cell.
Now I check if each data element can be converted to a float and
if so, convert it to a float. Now xlswriter will write the value
as a number event if it is not used in chart.