Add option to pass a 'prefix' string for sheet_name
Currently this system just takes all the *txt.tsv and *.png files
in a dir and creates an xlxs file in that dir.
If we are doing a scaling study then we might just have a perf
stat sheet and maybe an image file in each dir and I'd like
to create 1 xlsx file with the perf_stat and images from each
dir.
Excel has to have a a unique name for each sheet.
This option lets us pass a prefix string and all the sheets
(the *txt.tsv files really) will have that prefix in the sheet
name.