Allow for combining multiple dir's *txt.tsv files
Currently tsv_2_xlsx.py just reads all the *.txt.tsv files and
*.png's in a dir and creates a spreadsheet.
Say we are doing a scaling study of perf_stat data for requests
per second (RPS) at 1500, 1600, 1700, 1800 rps.
The script can easily create an xlsx file for each perf_stat
output.
But I want to compare 1500 vs 1600 vs 1700 vs 1800.
Opening each spreadsheet is a pain.
This commit allows us to pass a file containing the options to be
processed and 1 spreadsheet is created from the each set of
options.
So I can create 1 sheet with the perf_stat data from 1500-1800.
See the previous commit for an example script which does this.