save off lscpu info file for TSC freq
Currently 60secs doesn't save any identifying info about the cpu.
To properly process the perf_event data (in particular the %not
halted and <freq based on cyles) I need to know the TSC frequency.
I can compute it but, for now, I'll use lscpu Model name and
extract the freq from the Model name.
This could break easily. At intel prototype chips might not have
the TSC freq in the cpuid string.
I already have the code in spin.cpp to calculate the tsc freq
from a short loop and __rdtsc() but... I know the tsc is supposed
to be invariant (on recent cpus) but... I swear that on multi
socket boxes the tsc frequently is messed with over time by the
OS (I assume).
So I'll try this lscpu method until it doesn't work.