Add a log file with timestamps
Currently there isn't a good way to correlate the values with a
timestamp from some other system.
Add a 60secs.log file with the start time of each task.
Display both the local time (hardcoded to US/Pacific time) and an
epoch timestamp.
The MacBook doesn't support display fractions of a second of epoch
time so test for that and don't do ".%N" if not supported.