Add timestamps to the output so it can be correlated
Currently the perf stat is just charted without regard to time.
If everything is always constant interval then this is maybe okay.
But the interval might change and, if you want to correlate the
perf data with other charts, (maybe collected with a different
sampling interval or starting time) then you need a timestamp.
This commit uses the datetime stamp at the top of the perf stat
and constructs a beginning epoch timestamp.
Also now changed the charting python code to accept a column
number for the timestamp column and change from a line chart
to a scatter plot with points connected with staight lines.