Add timestamps to most tasks and plot pidstat over time
Currently none of the charts use a timestamp. I just plot the
data without timestamps. The data is just plotted in the order it
appears it in the sheet. This is okay but if you want to correlate
charts then it isn't reliable.
This change adds a timestamp to most sheets. I think the vmstat
table doesn't have any time in it so far.
Some of the tasks don't emit a starting time stamp. So I just use
the time I launched the task as the timestamp.
To make things easier for correlation, I use the timestamp of the
first task (see 60secs.log) as a global 'begin' time.
The charts are all relative to 'the number of seconds after the
global begin time'.
So I compute an epoch time for each row in the sheet. The I
subtract the global begin time to get a relative elapsed time.
This relative elapsed time is use as the X axis for the charts.
The type of charts have mostly been changed from 'line' to
scatter_straight (which means scatter plot with points connected
by straight lines). Also the relative timestamp column number is
now added to the hdrs line for the xlsx py script.

Currently the pidstat chart just did a bar chart of the average
process time for the top 40 processes.
Now I plot the top 20? processes %CPU over the whole run.
I also collect the IO per process but I haven't plotted that yet.

Lastly I added a script to get the requests per second and
response times and compute the java garbage collection times.