Add X axis, scatter plots, diff types charts/sheet
Currently none of the charts support a timestamp for the X axis.
The data is just plotted row by row.
If you collected data with a 1 second interval and the monitoring
tool can accurately keep up the 1 sec interval then the row number
can correspond to time.
But life is rarely so convenient.
This commit allows specifying a column for the X axis and using
a scatter plot so that irregular intervals are plotted accurately.