add script to get java garbage collection cpu usage
Initial commit of a java_gc_log_2_tsv.sh.
This scripts reads through the java GC log and pulls out the
cpu usage. It takes a begin time and end time and only emits
data rows for GC between the begin and end times.
It's not too pretty and I'm not a java expert so I don't know if
this actually represents all the time spent in garbage collecting.
This is java 1.8.
In awk, I pass the -1 2nd arg to mktime() to force it to get the
UTC time.