Add pidstat IO chart and option to size excel charts
Currently I'm collecting the pidstat IO per process data but I
don't plot it.
This commit adds a chart table to the tsv for the per process IO.
I also add a total IO.
I see some strangeness in the IO data. Some of it is impossibly
big. Like > 20 GB/sec. It makes me think pidstat is taking the
IO bytes and dividing it by the cputime for a process in that
interval. So I drop the crazy values.
The chart is still useful to see who is doing the IO.
I trust the iostat kBrd/s and kBwr/s much more.
Lastly, since I got tired of moving the charts around and resizing
them every time, I added an option to scale the size of the charts.
See the 'tsv_2_xlsx.py -s 2,2' option which scales the x,y by a
factor of 2 each.