Allow for changing size of charts and left align them
Currently the charts are the default size (5.5in by 3 in) and I
position them to right of the top row of the table.
This is annoying. The charts are too small and you have to scroll
down to the next table to see the next chart.
I added an option to specify an x,y scale factor for charts and
I move the chart to the left.
If there is more than 1 chart on a sheet I try to have them one
right below the previous chart.
On my macbook the default chart size seems to be about 15 rows
high. This probably varies by excel version, OS, maybe screen
resolution etc.
So if you have 3 charts starting at row 10, then the 1st chart
is at row 11, the next is at row 26, the next at 41.
If you specified a scale factor then the charts still start at
row 11, the 2nd at 11+2*15, the 3rd at 11+2*15+2*15.
You can specify the scale factor with -s x,y where x and y can be
floating point.
-s 1,1 means no scaling (the default).
-s 2,2 means double both the x and y size.
You can specify a 3 arg -s 2,2,20 where the 3rd number is the
number of rows that default (unscaled) chart takes.
The full default is -s 1,1,15