Add nicstat, flamegraph and more background support
Currently I use sar to monitor TCP and net interfaces.
The nicstat info seems to be more useful with %util and saturation
values, UDP and TCP.
I might drop the sar DEV and sar TCP collection for nicstat.
Also previously the data collections which were just copying files
every interval didn't support the background option.
Now, if you do the '-b -w' (run in background and wait) then, in
the wait loop, I do the copy tasks. So now I think everything can
be collected at once (for better or worse).
Lastly I wanted to be able to show what java is doing. The java
service is in a container so we have to go into the container and
start the sampling.
If you do the '-w' wait option then the flamegraph data collection
is started and stopped after the wait.
Currently I have the flamegraph profiler create a collapsed
stack file. It can create the svg but I want to feed the collapsed
file to a more recent flamegraph.pl.
I have scripts to put multiple svgs into one html file for
comparing.