Initial script for generating java flamegraphs
This uses https://github.com/jvm-profiling-tools/async-profiler
version 1.5.
I just generate a collapsed file because I want to feed it to a
more current flamegraph.pl. I have scripts to put multiple SVGs
in an html page.
This is an initial version which generates flamegraph data for
java in a container. It has hard coded to look for a 'gmatch'
string in the java command.