Improve support for combining xlsx files
I want to be able to take the sheets from 2 directories and make
1 xlsx file with each of the sheets from the directories.
The prefix string makes the sheet names unique.
The code sorts the file names and puts sheets (with the same base
filename) next to each other.
So if we have dirs 1300rps and 1500rps and both have a
sys_01_uptime file then the resulting xlsx would have a sheet
named 'uptime_1300rps' and uptime_1500rps.
I'd really like to have the data for both .tsv files on the same
page. With the 2 sheet solution I have to click back and forth.