This script read 1 or more svgs and creates an html file
If we have multiple svgs, I'd like to be able to have them all
on the same page. This script will read multiple svgs and put
them on the same html file.
Unfortunately the svg has stuff that seems to be globally defined
variables and functions so I have to rename them.
This is intended for things like req/sec scaling studies.
Also, for the case of 2 svgs, you can put one regular svg and
create the 2nd svg inverted. Then when you combine them with
svg_to_html.sh the 2 bases will be next to each other.
Lastly this script lets you pass a 'factor' (like 1500 for 1500
req/sec) and the string for each svg box gets a 'samples/metric= x'
string where x is sample_count/factor.
This is intended to make it easier to compare 2 svgs.