Add -t allx or ally or allxy option and chip specific events
The perf record sched_switch task can generate lots of data and
generating flamegraphs depends on knowing details of the java
service running on the server.
Change the code so that sched_switch and/or flamegraphs are only
selected if you explicitly ask for them (with -t allx or -t ally
or -t allxy). You can also explicitly select them with
-t sched_switch or -t flamegraph.
Some of the perf_stat events are cpu specific (currently setup for
broadwell servers).
Added a decode_intel_fam_mod.sh to decode the cpu type.
Changed the perf stat task to have broadwell, haswell, skylake and
cascade lake specific events. This is not so bad since most of the
cpu-specific events are memory traffic events and they don't
change that much between these skus.
Also added a '60secs started' (with timestamp) message and a
'60secs ended' (with a timestamp) message to a log file.