Add option for prefix output dir with a date/time string
Currently run_60secs.sh creates an output dir with a date time
string prefixed. Like 20-02-24_155959_60secs.
This can be useful but it also complicates post-processing.
This commit makes adding the prefix optional.
Use the -P option to add the prefix.
Also, there is a bit of code to find the gmat docker container iD.
This container doesn't exist on all machines.
If the container isn't found then don't pass it to 60secs with
-C containerID