Allow -t all -a flamegraph,sched_switch options
The flamegraph and sched_switch options have extra overhead.
I'd still like to be able to specify '-t all' and not get
flamegraphs and/or sched_switch tracing by default.
So for the moment going with -t all -a flamegraph|-a sched_switch
Also I had a CPUname wrong (Braswell instead of Broadwell).