Add multiple charts per data table and timestamps for all
Currently vmstat, mpstat and iostat charts are just line charts.
Line charts just plot the data in row order. This is probably
okay most of the time if the interval is 1 sec but there is a time
given in the output of each collection.
This commit converts the data+time in the data to an epoch time
and, given an input "begin the data collection" epoch timestamp,
the data is plotted relative to this beginning timestamp.
Now all the charts are on the same timescale except vmstat.
Also, currently many charts have line with very different yrange
data values. For instance the vmstat chart plots billions of bytes
in 'cache' and %idle (which varies between 0-100).
I added the ability to specify which columns to be added to a
chart.
This lets me, continuing with the vmstat example, put all the
memory lines on 1 chart, all the cpu related lines on 'vmstat cpu'
chart, the block io on another chart, the swap pages on another,
etc.
Now I don't have to keep deleting the top lines (which causes
excel to rescale the chart to the remaining lines) in order to see
the lower lines.
Also, for vmstat, I renamed the headers to make them more
meaningful to new users (so 'r' column -> 'runnable', etc).