Add debug info for add chart errors
Currently this py script works fine if you actually have a data
table to plot.
But there are lots of reasons why the data table might be empty.
Say you restricted the selecting of the RPS data or response time
to a particular time range. And the system was idle during that
time (so no RPS).
So just print a message that there is no data for a sheet/file
and continue (skipping trying to add the chart... which would
fail anyway).