Workaround way too big pidstat %CPU values
Don't know why but pidstat sometimes gives a value for %CPU for
a process and then a really bit value... like 1000x too big.
Not sure what I should do here. It seems to coincide with lots of
IO by other processes.
But... as a hack, just divide the values by 1000.0. If they are
still bigger than num_cpus*100.0 then set them to 0.0.