Add top 40 hot functions chart from java.collapsed
As helpful as the flamegraphs are, it can be hard to see where
the time is going sometimes.
The info is all there in java.collapsed.
This script reads the java.collapsed file and creates a top 40
hot function table and a column chart for the xlsx.
Just the last function in each call stack is hashed (the call
stack is split on ';').
The top 40 routines are shown with a %of_total_samples and a
%cumu_total_samples column for each function.