Add request latency data support
Currently this file handles the json data for RPS over time and
the p50/p95/p99 request response times.
This commit adds another breakout of response times. The data is
for each interval, the number of responses in a given time bucket
like 1ms-2ms, 2-3ms, etc.
This lets me better see how the response times are distributed.
Plot a line chart of response times for each bucket over the
interval of the run.
Also, display a chart showing the %total time for each bucket
where the total time for a bucket is the 1st number of the bucket
string (ie 1 for 1ms-2ms) multiplied by the total number of
responses in that bucket. So the 1 is the lower bound for the
time bucket and the 2 is the upper bound.
It would be better if I had a average response time for each
bucket but I don't have that. If I did have it I might be able to
compare the total_time to the java cpu time. As it is now the
java cpu time is between:
 lwr_bnd*num_responses <= cpu_time <= upr_bnd*num_responses