Add support for summary file stats
Currently the raw data is charted. That's good.
But I frequently have to do an average then of the data over time
to see the average memory bw or qpi bw etc.
So add support for creating summary stats (just averages for now)
of selected columns of data.
For perf_stat_scatter, this means accepting as an arg the name of
the summary_file.
And the fields to be averaged are passed on the awk cmdline.