Add support for summary file stats
Currently the data is plotted for each column. That's good.
But frequently I want to average a field over the run.
Like %idle from vmstat or RdKB/sec from nicstat.
This commit adds a 'summary stat' feature for selected perf data.
The data is written to a summary file.
The fields to be averaged are passed on the awk cmdline.
Since the column names are frequently kind of opaque, allow for
passing a column name to match on and a string to use instead of
the column name.
Also, some of the fields are 'value'/second. For the 'per second'
fields, recompute the 'value' (value/second * interval),
accumulate the value and divide it by the elapsed time.
This gives a more accurate average over time.
Also, I want 1 average to print... so for resources like disks
where you may have more than 1 disk, I sum all the IO and divide
it by the elapsed time. If you have 1 'rotating' disk and 1 SSD,
and you are writing just to the slow spinning disk, that might be
maxed out and you might have a hard time seeing it from the
summary stats.
I haven't added a summary page to the xlsx file yet.
This is mostly (first) for the "run on a bunch of on-prem servers
and compare the results" case.