Try changing the chart_style property
When I import the xlsx file into google sheets the charts aren't
so good.
For instance, I use the excel scatter plot with straight lines
connecting the data points and no markers. But google sheets
changes this to a scatter plot with no lines and adds markers.
I wanted to try and see if changing the chart style help sheets
plot stuff properly. But it doesn't at least for the limited
testing I did.
I tried changing the ch_style to 14 for scatter_straight plots
and it just changes the color pallette in excel.