Example script to generate xlsx file from 60secs data files
Currently input to the script is just the dir name of 60secs data
files.
The script will read each handled type of data file and create a
.tsv file (in the same data dir) for input to the python script
which will create an Excel xlsx file with charts.
The xlsx file will be in the same dir.
The script tries to extract an 'rps' (requests per second) number
from the dir path (for example I might have ._1200rps_... in the
path name to identify the load placed on the server).
This rps factor is added to the sheet names in the xlsx.
If more than 1 dir is passed to the script then a summary xlsx is
created with all the sheets from each dir. The rps factor on the
sheet name then keeps the sheets from having the same name.