Read phase, ipmitool sdr data and summary table
Read the ipmitool sdr power data and convert it to a table.
I wanted to create a summary table of the most important data
variables across tables.
So I added a 'compute average' for various sys*.tsv files.
This table is added as the first sheet in the spreadsheet.
Also, if there is a phase file and if a table has a timestamp
column then the phase for each line is added.
The phase file has the format:
benchmark1 beg_tm1 end_tm1
benchmark2 beg_tm2 end_tm2
etc
Also fixed errors in computing epoch time (the ampm variable name
was mistyped on the  dt_to_epoch() call).