Add support for phase file
The phase file format is:
benchmark0 beg_tm end_tm
benchmark1 beg_tm end_tm
For each line of a tsv file, if it has a timestamp, then the
timestamp is checked to see if it belongs to a phase and if so,
the phase is added to the 'phase' column.
This lets you do a data filter and select the lines for the
phase.