Change resp_2_tsv.sh call to use getopts and toplev flamegraphs
Now resp_2_tsv.sh uses options (not positional args) so change the
calls to use options.
I've changed the toplev processing to do a flamegraph now.
The toplev data is collected per core and now I roll it up to a
a server level. I'd like to be able to roll up multiple servers
into the flamegraph but I haven't tested that aspect yet.