Add script to generate toplev flamegraph
This script reads the toplev per core data and combines it to the
server level and generates a collapsed call stack type output file
suitable for generating a flamegraph.
The toplev info is generated with the command (from 60secs.sh)
          sysctl kernel.nmi_watchdog=0 && export PERF=$PERF_BIN && python ${SCR_DIR}/pmu-tools-master/toplev.py -l3  -x, --no-multiplex  -o $FL.csv -v --per-core --nodes +CPU_Utilization  -- sleep $INTRVL
The cmd does about 20 pmu collections to get all the data.
The CPU_Utilization is collected so that I can show the 'halted'
value for the server.