Add stdev option for summary stats
Currently the summary data only does the average of a variable
over the interval.
It is helpful to also produce stddev (and/or %stdev) for a
variable.
Add a 3rd field {||stdev} to the sum_flds option for a variable.
Only some of the sys files have this feature added.