Add script to collect power via ipmitool
This script collects power for a server using the ipmitool.
It sometimes call other ipmi based tools which handle user ids
and passwords.
I've tried various APIs. The max and average over an interval does
not seem to work, at least not to me.
So I get the instantaneous value and compute an average and max.