Start cross metrics (like instructions/request)
Currently the summary sheet has a line for each metric added to
the summary file.
For doing scaling studies such as results at different frequencies
or diffent request per second (RPS) it is usefull to compute cross
metrics such as instructions/request.
This commit reads the sum.tsv file, looks for the RPS metric and
adds a new column 'value/1000_reuests' which divides any 'per
second' metric (looks for /s) and divides it by RPS/1000.
This gives an easier way to check if things scale well as requests
per second are ramped up.