Add support for cascade lake QPI bw and missing events
I added cascade lake QPI (UPI) bandwidth events and events to get
memory bandwidth for chips with more memory channels (like cascade
lake).
The PMC placeholders are in /sys/devices and appear like
uncore_upi0, uncore_upi1, uncore_upi2.
If the 60secs script sees the sys/devices PMC exists then it adds
the events for that PMC.
The problem for this script is that perf may not collect/print the
event even though the sys/devices PMC exists.
This script will compute new metrics based on events which may or
may not appear in the input perf stat file.
Currently all the dependent events have to appear in the input
file. If none of the dependent events are in the input file then
that is okay... the metric is just not computed.
If this script has a list of events in the got_lkup lists and only
some of the events appear in the input event data then the headers
might not match up with data in the column.
This commit relaxes the requirements for the 'some of the events
occurs'. It also fixes the case where the lookup for a missing
event might not show that the event wan't found (and so it will
reference the wrong event).