handle timezone diff and add gmatching log charts
Currently the timestamps on different files might not be for the
same timezone. Some files use the epoch seconds in UTC, some files
just print a date/time stamp with a timezone stamp.
This commit adds a step which computes the difference between the
localtime date/time stamp and UTC time. And this difference is
passed to routines which need it.
This commit also adds support for gmatching service log files.
This is a very specific file. I wanted to correlate msgs in the
log files with delays due to lock contention and errors in request
servicing.
Lastly, if the script sees an async-profiler collapsed+traces
callstack file then it creates a top 100 function hotspot table
for the traces data.