Add 'watch' work type and cascade lake perf events
It can be useful to execute a command periodically, gather the
output, post-process the output  and plot it in a chart.
The 'watch' option lets a user enter a command to be executed
periodically, collect the output, and post process it.
The current 'post processing' allows for processing the
'ls -l filename' output.
Also, the perf cascade lake events don't have qpi bw events.
This commit adds the qpi(upi) events.