Add support for summarizing many servers
Currently 1 excel xlsx file is create per results directory.
Or if we have more than 1 results dir, an output xlsx file is
created with all the sheets from the results dirs in the xlsx
file.
This is not workable for hundreds of servers.
This commit will create average the perf_stat data files from each
dir and chart it.
Here is a cmdline to process 350 servers' results dirs and
create an xlsx file, cutting off values > 1000000 and -S skipping
creating the xlsx file in each results dir and -A do averaging.
perf will sometimes create bogus very, very big values and the
-m max_val option lets you set these big values to 0.
~/repos/60secs/gen_xlsx.sh -d archive_itp_200614_350_11h_10sec -X xlsx/archive_itp_200614_350_11h_10sec_tst1 -m 1000000  -S  -A