add support for averaging many sheets into 1 sheet
Currently the logic for handling multiple results dirs is king of
muddled.
This commit improves the code.
The options for each results dir are put into an input options
file.
Then a hash is made of all the types of files in the results dirs.
A loop over each base file name type is done and then a loop over
each results dir.
This groups all the same base name results together in the xlsx
file.
For the averaging multiple results dir into 1 sheet, we sum the
values of each cell and accumulate a count of how many values are
being accumulated.