move the tsv lines down 40 lines for charts
Currently the charts are all put one below the other at the left
side of the sheet.
This can be okay but if you filter the data in the sheet then the
charts can disappear if the rows under the chart are filtered out.
This commit moves the data lines lower and, if tsv_2_xlsx.py sees
that the first 35 or so lines are blank then the charts are put
across the top of the sheet.