Move metric_out lines down and create phase from cpu2017 data
Moving the lines down (35 lines down) in the tsv tells the python
script to put the charts horizontally across the top as opposed to
vertically down the left side.
This avoids hiding the charts if you filter (hide) some rows.
Also, for specint, create a phase file from the CPU2017.001.log
file.