Add support for clipping data to a phase and debug option
Currently if there is a phase file, the phase is just included as
column 0 of the sheet. This is useful but the summary stats do an
average over all the data in selected columns.
This commit improves on this by letting the cmdine specify to only
show rows for intervals in the phase file.
Option -c y accomplishes this.
The script just blanks out the rows outside a phase. This ensures
that the rows aren't included in the average.
Also to make the summarizing more flexible, an average of each
column is done above the header. This average only includes the
data in the clipped rows (the rows with the phase non-blank).
Then the summary sheet does an indirect address reference for
each column's header and average.
With this change, if you filter the rows on some condition then
the summary average for the affected columns will also reflect the
filtering.