Add debug option to exit early and formulas for summary stats
Currently the processing of the perf_stat data can take a couple
of seconds per file. This is not too bad but if you have say 350
servers you want to combine the stats for, this is 10+minutes.
In particular, many times I need to regenerate a spreadsheet but
the perf_stat tsv file is fine and doesn't need to be regenerated.
So if you do '-D skip_perf_stat_scatter ' as an option for
gen_xlsx.sh then the perf_stat_scatter.sh script just exits.
Also I wanted to make the summary stats more flexible.
Currently the averages in the summary sheet are computed by the
script and written 'hard-coded' to the summary sheet.
If I do a filter on the data, the summary averages don't change.
Fix this by setting a column average (above the header row) and
use the formula to make the average omit filtered out rows.
Then make the summary sheet reference the column average and
header.