Fix a type skt_inc, fix that new eqn reqirements filled
A reference skt_inc should be skt_incr.
Also, one of the TMAM new equations has a faulty check to see if
the required events are in the input file.
Fix the skt_incr typo. This fixes handling of perf stat data by
socket.
Also, the TMAM equation TMAM_Backend_Bound_BadSpec uses the
'formula' type and all the required events need to be in the input
file. This commit fixes the code so that the equation is not
output if all the required input events aren't present.