Reverse key and value pair in sum_file table
Currently the key text is printed first then the value is printed.
This is okay but the key text can be obscured in the excel cell by
the value text.
Printing the value first then the key lets the key text flow into
the adjacent cells (if they are empty) and the text isn't
obscured.