Reduce verbosity and start file with llist of input files
Reduce volume printed out especially for the case where we have
lots of input dirs.
Currently the list of input files (directories) has to be built by
gen_xlsx.sh.
I'd like to be able to have more flexibility for the list of dirs.
Say to be able to subset the directories by make/model of server.
Or process just the summary files.
The new -I input_file option is the start of this new feature.