Add support for -I input_file_list
Currently you can select a dir for processing with -d dir_name.
If there are multiple data dirs under dir_name then they are all
added to the list of sub-dirs to be processed.
Let say I want to create a custom list of sub-dirs.
In particular say I want to process all the sum.tsv summary files
under a dir (and not the other data files). This assumes that the
sum.tsv files have already been created.
So create the list of sum.tsv files:
  find archive_itp_200626_350_full_itp -name sum.tsv > t.lst
Then process the list with:
  ~/repos/60secs/gen_xlsx.sh -I t.lst -X xlsx/tmp
Then I can do (on macbook):
  open xlsx/tmp_all.xlsx