Lots of changes to speed up processing
I want to be able to process 100's of server data files and
produced 1 xlsx file for say, the day's data.
And I want to average the 100s of server files into 1 variable
representing the average across all the servers. Like compute
average memory bw across the 100s of servers.
And then I want to be to do multiple days.
And then I want to be able to put the multiple days of data/charts
in 1 xlsx so it can be easily compared.
Probably the main speed impprovement is to launch up to 12
sys_2_tsv.sh scripts in the background at a time. When I get to
12 I wait for the jobs to finish then continue.
So I average the results of the 100s of servers to get a cluster
level average. If I want to combine these 'average' charts into
a 'multi-day' charts file, then I have to save off the averaged
data.
the -a avg_dir option is where I save the averaged data files.
I also want to put all the charts for one day in a row, then
the charts for the next day in a new row, etc.
This makes it easy to compare different day's charts.
I use the -F desc_file option to indicate that a set of files
belong to the same day. If the name of the desc_file changes
to a new file name, then I start the charts for that group on a
new row. Also I print the contents of the desc_file at the top
left of the first chart so you can ID to whom the charts on that
row belong.
Also, the muttley data (like RPS over time) is usually collected
at 10 second intervals. I might collect perf data at 30 sec
interval (for a 14 hour run). And, if you want to divided the perf
data by the RPS, you need to RPS data to have the same number of
rows as the perf data. So I added an option
match_itp_muttley_interval. If I have muttley data and perf stat
(itp) data then I grab the perf stat interval and compare it to
the muttley interval. If the perf stat interval (say 30secs) is
greater than the muttley interval (say 10 secs) then I only use
every 3rd (30/10) muttley data record. I could get better and
do the average over the 3 records but this is sort of a hack
anyway.
Previous changes messed up the sum.tsv handling so I fixed that I
hope.