Add support for -a avg_dir and averaged tsv files
In order to save time, I wanted to be able to save off the
averaged tsv data. Say I average 100s of servers to get an average
value for memory bw. If I want to make an xlsx file with just the
averages, I don't want to have to recompute the 100s of files.
These changes are part of saving off the averaged tsv files.