Add average column to sum.tsv file
Currently if there are multiple values for a metric on the summary
sheet (due to having multiple files being included in the sheet)
then the average of each metric is listed.
For example, if we have 400 servers and the average Mem_bandwidth
for each server, then the summary sheet will show 400 columns of
mem_bandwidth (one for each server).
In that case I'll always want the average.
So compute an average and include that as the 1st column.