Support multiple specint bmarks and date str to sum.tsv
Currently for CPU2017 specint rate logs, only 1 benchmark is
supported. But we might have multiple (different) subtests in the
same run.
Allow multiple subtests in the same run. The score, runtime,
number of copies is added to the sum.tsv file and prefixed to the
chart titles.

Currently the date/time strings are not put in the sum.tsv. Having
the begin/end time (in human format) can be useful for things like
verifying that the data is from the run you expected or going to
grafana to see that the RPS was for that time interval.
This is a partial implementation of having the date/time info.
I say partial because it just looks in the run.log file (in the
same dir as the sys*_perf_stat.txt file) and grabs the begin/end
timestamps from there.

Also, the instruction throughput (Instructions/sec) is calculated
the correct way.
Now calculate it based on:
num_cpus*0.01*%cpu_utilization*freq/cpi

Also, the muttley RPS info doesn't appear in the sum.tsv at all.
Add a min and max stat for chart added to the sum.tsv. This is for
the cases where I'm monitoring say 14 hours of load. An average
isn't so helpful.
I probably should add an average too for cases where the interval
isn't as long and the average represents the load well.