Allow running python in background and track load
Currently the calls to sys_2_tsv.sh are started in the background.
When the number of jobs started are greater than the BACKGROUND
variable then the script does a wait till all the jobs are done.
This doesn't take any account of the load on the system.
Also, after all the sys_2_tsv.sh calls are done then the python
call to tsv_2_xlsx.sh is done synchronously and single threaded.
It would be good if, when we are generated lots of spreadsheets,
to be able to put the python call in the background too.
The python script can take a while though.
So we need to track the load and make sure we aren't overloading
the system.
We use the uptime cmd to track the load.
The python script is the last thing done in sys_2_tsv.sh. So we
write the pid of the python background task to a 'sys_2_tsv.pid'
file. Then the script calling gen_xlsx.sh has to check if
the python scripts are done.