disable matching perf & json sampling period if needed
Currently, if the match_itp_muttley_interval option is specified
then this script tries to drop muttley samples until the muttley
samples/interval is the same as the itp samples/interval.
This works okay if the muttley interval is the usual 10 secs and
the ITP interval is a multiple of 10 seconds.
In some cases the muttley interval is > than the itp interval.
In this case then this script won't try to match the intervals.