Fix the Frontend_Bound% and Retiring% formula
The frontend_bound, %retiring and backend_bound calculations are
not working. The missing piece is the multiply frontend_bound and
retiring by 'cycles/cpu_clock_unhalted.thread_any'. This factor
is == ~0.5 if you are running on just one of the HT threads of
each core. If you are running on both HT threads of a core then
the factor is == 1.0. So the factor tells us how much of the
cycles are using both of the HT threads.
Now the formulas are pretty much identical to Intel ITP result.
Also add all computed metrics to the summary sheet if the metrics
aren't already added to the summary.