Read Intel ITP TMAM data and gen flamegraph callstack
The Intel ITP utility generates a toplev/TMAM report such as in
the metric_out.average file.
The TMAM info has the %retiring, %frontend_bound, %backend_bound,
%bad_speculation metrics (for level 1).
ITP then breaks down the each category into sub categories and
goes 4 levels deep.
This data can be converted to a callstack-like csv file and use
flamegraph.pl to generate a svg.
This script will read the metric_out.average file and create a
csv suitable for input to flamegraph.pl.
Flamegraph.pl will generate a svg from the csv.
Several 'hmmm' things are done to convert the data.
I assume that the sum of the child row should sum to <= the parent
node. Usually this is okay. The port_utilization child row seems
to sum to 100% even though the %port_utlization parent node might
only be 10%. So the child row sums to 100% but the parent is 10%.
My conversion rescales the port_utilization child row values to by
parent_value/child_sum.
I do this scaling for any parent+child_row where the child sum is
greater than the parent.
Also, since the flamegraph call stack format propagates the child
layer values to the parent.
So if you have
A 100
  B 20
  C 50  meaning 100% of the tie is psnt in A and, further, 20%
of A's time is spent in B and 50% in C.
For flamegraph this needs to be like
A 30
A;B 20
A;C 50