Reduce chart title text for ITP and add flamegraph
Currently if you are doing specint and have multiple scores in a
directory tree then the scores are added to the chart title for
ITP charts. The current added text repeats the subtest name and
number of threads.
Change to only put the subtest name and copies once.
Also, if you have ITP TMAM data, then create a flamegraph of the
TMAM data and include a PNG of the flamegraph in the excel sheet.